Initial cut flattening out column directory
